Optimization

+++ app/Http/Controllers/API/DepartmentController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\API;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller as Controller;
-use App\Http\Controllers\API\BaseController as BaseController;
+use App\Http\Controllers\BaseController as BaseController;
 use App\Department;
 use Validator;
 
@@ -17,12 +17,12 @@ class DepartmentController extends BaseController
      */
     public function index()
     {
-       $departments = Department::all();
+     $departments = Department::all();
 
-       return $this->sendResponse($departments->toArray(), 'Departments retrieved successfully.');
-   }
+     return $this->sendResponse($departments->toArray(), 'Departments retrieved successfully.');
+ }
 
-    
+ 
 
     /**
      * Store a newly created resource in storage.
@@ -30,21 +30,22 @@ class DepartmentController extends BaseController
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
-    {
-         $input = $request->all();
+    // public function store(Request $request)
+    // {
+    //      $input = $request->all();
 
-       $validator = Validator::make($input, [
-        'name' => 'required|string|max:60|unique:departments',
-        
-    ]);
+    //    $validator = Validator::make($input, [
+    //     'name' => 'required|string|max:60|unique:departments',
+    
+    // ]);
 
-       if($validator->fails()){
-        return $this->sendError('Validation Error.', $validator->errors());       
-    }
-    $department = Department::create($input);
-    return $this->sendResponse($department->toArray(), 'Department created successfully.');
-    }
+    //    if($validator->fails()){
+    //     return $this->sendError('Validation Error.', $validator->errors());       
+    // }
+    // $department = Department::create($input);
+    
+    // return $this->sendResponse($department->toArray(), 'Department created successfully.');
+    // }
 
     /**
      * Display the specified resource.
@@ -54,16 +55,15 @@ class DepartmentController extends BaseController
      */
     public function show($id)
     {
-     $department = Department::find($id);
+       $department = Department::find($id);
 
-     if (is_null($department)) {
+       if (is_null($department)) {
         return $this->sendError('Department not found.');
     }
-
     return $this->sendResponse($department->toArray(), 'Department retrieved successfully.');
 }
 
-    
+
 
     /**
      * Update the specified resource in storage.
@@ -74,20 +74,26 @@ class DepartmentController extends BaseController
      */
     public function update(Request $request, Department $department)
     {
-        $input = $request->all();
+        try {
+         
+            $input = $request->all();
 
-        $validator = Validator::make($input, [
-            'name' => 'required|string|max:60|unique:departments',
-        ]);
+            $validator = Validator::make($input, [
+                'name' => 'required|string|max:60|unique:departments',
+            ]);
 
-        if($validator->fails()){
-            return $this->sendError('Validation Error.', $validator->errors());       
-        }
-        $department->name = $input['name'];
-        $department->save();
+            if($validator->fails()){
+                return $this->sendError('Validation Error.', $validator->errors());       
+            }
+            $department->name = $input['name'];
+            $department->save();
 
-        return $this->sendResponse($department->toArray(), 'Department updated successfully.');
-    }
+            return $this->sendResponse($department->toArray(), 'Department updated successfully.');
+
+        } catch (Exception $e) {
+          return $this->sendError('Department delete Unsuccessful.', $e->getMessage());  
+      }
+  }
 
 
     /**
@@ -96,10 +102,22 @@ class DepartmentController extends BaseController
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Department $department)
+    public function destroy( $id)
     {
-        $department->delete();
+        try {
+            $department = Department::find($id);
 
-        return $this->sendResponse($department->toArray(), 'Department deleted successfully.');
+            if($department != null){
+              $department->delete();
+              
+              return $this->sendDelete('Department delete Successful.');
+          }else{
+            return $this->sendDelete('Department not found. Enter the Valid Deprtment');
+        }
+
+    } catch (Exception $e) {
+        return $this->sendError('Department delete Unsuccessful.', $e->getMessage());
     }
+    
+}
 }
