+++ app/Http/Controllers/API/EmployeeController.php
@@ -4,11 +4,11 @@ namespace App\Http\Controllers\API;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller as Controller;
-use App\Http\Controllers\API\BaseController as BaseController;
+use App\Http\Controllers\BaseController as BaseController;
 use App\Employee;
 use App\Department;
-use APIErrorValidationResponse;
 use Validator;
+use Illuminate\Support\Facades\DB;
 
 
 class EmployeeController extends BaseController
@@ -20,20 +20,16 @@ class EmployeeController extends BaseController
      */
     public function index()
     {
-       $employees = Employee::all();
-       return $this->sendResponse(json_decode($employees), 'Employees retrieved successfully.');
-   }
+        $employees = DB::table('employees')
+        ->leftJoin('departments', 'employees.department_id', '=', 'departments.id')
+        ->select('employees.*', 'departments.name as department_name', 'departments.id as department_id')
+        ->orderBy('employees.id','ASC')
+        ->get();
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-      
-}
+        return $this->sendResponse(json_decode($employees), 'Employees retrieved successfully.');
+    }
 
+    
     /**
      * Store a newly created resource in storage.
      *
@@ -42,20 +38,19 @@ class EmployeeController extends BaseController
      */
     public function store(Request $request)
     {
-         $input = $request->all();
-
+     $input = $request->all();
          // $department = Department::all();
-       
-          $messages = [
-    'name.required' => 'The :attribute is required',
-    'add.required' => 'The :attribute is required',
-    'birthdate.required' => 'The :attribute is required',
-    'date_hired.required' => 'The :attribute is required',
-    'department_id.required' => 'The :attribute is required',
 
-];
+     $messages = [
+        'name.required' => 'The :attribute is required',
+        'add.required' => 'The :attribute is required',
+        'birthdate.required' => 'The :attribute is required',
+        'date_hired.required' => 'The :attribute is required',
+        'department_id.required' => 'The :attribute is required',
 
-       $validator = Validator::make($input, [
+    ];
+
+    $validator = Validator::make($input, [
         'name' => 'required|string|max:10',
         'add' => 'required|string|max:120',
         'birthdate' => 'required|date',
@@ -64,13 +59,13 @@ class EmployeeController extends BaseController
 
     ], $messages);
 
-       if($validator->fails()){
+    if($validator->fails()){
          //return new APIErrorValidationResponse($request, $validator);
         return $this->sendError('Validation Error or Invalid Json Format.', $validator->errors());       
     }
-    $department = Employee::create($input);
-    return $this->sendResponse($department->toArray(), 'Employee created successfully.');
-    }
+    $employee = Employee::create($input);
+    return $this->sendResponse($employee->toArray(), 'Employee created successfully.');
+}
 
     /**
      * Display the specified resource.
@@ -80,7 +75,12 @@ class EmployeeController extends BaseController
      */
     public function show($id)
     {
-        $employee = Employee::find($id);
+        //$employee = Employee::find($id);
+        $employee = DB::table('employees')
+        ->leftJoin('departments', 'employees.department_id', '=', 'departments.id')
+        ->select('employees.*', 'departments.name as department_name', 'departments.id as department_id' )
+        ->where('employees.id', '=', $id)
+        ->get();
 
         if (is_null($employee)) {
             return $this->sendError('Employee not found.');
@@ -89,17 +89,7 @@ class EmployeeController extends BaseController
         return $this->sendResponse($employee->toArray(), 'Employee retrieved successfully.');
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
-    }
-
+    
     /**
      * Update the specified resource in storage.
      *
@@ -109,40 +99,40 @@ class EmployeeController extends BaseController
      */
     public function update(Request $request, Employee $employee)
     {
-        $input = $request->all();
+        try {
 
-           $messages = [
-    'name.required' => 'The name:attribute is required',
-    'add.required' => 'The add:attribute is required',
-    'birthdate.required' => 'The birthdate:attribute is required',
-    'date_hired.required' => 'The date_hired:attribute is required',
-    'department_id.required' => 'The department_id:attribute is required',
-    
-    'name.json' => 'The name :attribute field must be valid JSON',
-    'add .json' => 'The :add attribute field must be valid JSON',
-    'birthdate.json' => 'The birthdate :attribute field must be valid JSON',
-    'date_hired.json' => 'The date_hired :attribute field must be valid JSON',
-    'department_id.json' => 'The department_id :attribute field must be valid JSON',
-];
-        $validator = Validator::make($input, [
-            'name' => 'required|string|max:10',
-            'add' => 'required|string|max:120',
-            'birthdate' => 'required|date',
-            'date_hired' => 'required|date',
-            'department_id' => 'required|exists:departments,id',
-        ],$messages);
+            $input = $request->all();
+            $messages = [
+                'name.required' => 'The name:attribute is required',
+                'add.required' => 'The add:attribute is required',
+                'birthdate.required' => 'The birthdate:attribute is required',
+                'date_hired.required' => 'The date_hired:attribute is required',
+                'department_id.required' => 'The department_id:attribute is required',
 
-        if($validator->fails()){
-            return $this->sendError('Validation Error or Invalid Json Format.', $validator->errors());       
+            ];
+            $validator = Validator::make($input, [
+                'name' => 'required|string|max:10',
+                'add' => 'required|string|max:120',
+                'birthdate' => 'required|date',
+                'date_hired' => 'required|date',
+                'department_id' => 'required|exists:departments,id',
+            ],$messages);
+
+            if($validator->fails()){
+                return $this->sendError('Validation Error or Invalid Json Format.', $validator->errors());       
+            }
+            $employee->name = $input['name'];
+            $employee->add = $input['add'];
+            $employee->birthdate = $input['birthdate'];
+            $employee->date_hired = $input['date_hired'];
+            $employee->department_id = $input['department_id'];
+            $employee->save();
+
+            return $this->sendResponse($employee->toArray(), 'Employee updated successfully.');
+        } catch (Exception $e) {
+            return $this->sendError('Employee Update Operation Failed',$e->getMessage());
+            
         }
-        $employee->name = $input['name'];
-        $employee->add = $input['add'];
-        $employee->birthdate = $input['birthdate'];
-        $employee->date_hired = $input['date_hired'];
-        $employee->department_id = $input['department_id'];
-        $employee->save();
-
-        return $this->sendResponse($employee->toArray(), 'Employee updated successfully.');
     }
 
     /**
@@ -151,10 +141,22 @@ class EmployeeController extends BaseController
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Employee $employee)
+     public function destroy( $id)
     {
-        $employee->delete();
+        try {
+            $employee = Employee::find($id);
 
-        return $this->sendResponse($employee->toArray(), 'Employee deleted successfully.');
+            if($employee != null){
+              $employee->delete();
+              
+              return $this->sendDelete('Employee delete Successful.');
+          }else{
+            return $this->sendDelete('Employee not found. Enter the Valid Deprtment');
+        }
+
+    } catch (Exception $e) {
+        return $this->sendError('Employee delete Unsuccessful.', $e->getMessage());
     }
+    
+}
 }
